Change encryption, Test UDP and HTTP on vmware

+++ BarbaTunnel/BarbaTunnel/BarbaUtils.h
@@ -31,7 +31,7 @@ public:
 	static std::tstring GetFileFolderFromUrl(LPCTSTR url);
 	static std::tstring GetFileUrlFromHttpRequest(LPCTSTR httpRequest);
 	static std::tstring GetKeyValueFromString(LPCTSTR httpRequest, LPCTSTR key);
-	static u_int GetKeyValueFromString(LPCTSTR httpRequest, LPCTSTR key, u_int defValue);
+	static size_t GetKeyValueFromString(LPCTSTR httpRequest, LPCTSTR key, size_t defValue);
 	static std::tstring FormatTimeForHttp();
 	static std::tstring FormatTimeForHttp(time_t* t);
 	static std::tstring ConvertIpToString(u_int ip);
